requirements.txt
streamlit
pandas
selenium
webdriver-manager
gspread
google-auth
plotly
undetected-chromedriver
